+++ pytorch_lightning/trainer/evaluation_loop.py
@@ -515,7 +515,7 @@ class TrainerEvaluationLoopMixin(ABC):
 
         # enable fast_dev_run without val loop
         if dataloaders is None:
-            return
+            return [], []
 
         # cap max batches to 1 when using fast_dev_run
         if self.fast_dev_run: